Создание структуры. Создание компонента Card.
