+++ src/app/app.tsx
@@ -1,4 +1,4 @@
-import Main from '../main/main';
+import Main from '../pages/main/main';
 
 type AppCountArendaProps = {
   countArenda : number;
