+++ src/components/card.tsx
@@ -1,80 +1,8 @@
 
-const directions:string[] = ['Paris','Cologne','Brussels','Amsterdam','Hamburg','Dusseldorf'];
-type CountArendaProps = {
-  countArenda : number;
-}
-
-export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
+export default function Card() : JSX.Element {
   return (
-    <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link header__logo-link--active">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
-      <main className="page__main page__main--index">
-        <h1 className="visually-hidden">Cities</h1>
-        <div className="tabs">
-          <section className="locations container">
-            <ul className="locations__list tabs__list">
-              {
-                directions.map((item) => (
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item" href="#">
-                      <span>{item}</span>
-                    </a>
-                  </li>)
-                )
-              }
-            </ul>
-          </section>
-        </div>
-        <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{countArenda} places to stay in Amsterdam</b>
-              <form className="places__sorting" action="#" method="get">
-                <span className="places__sorting-caption">Sort by</span>
-                <span className="places__sorting-type" tabIndex={0}>
-                  Popular
-                  <svg className="places__sorting-arrow" width="7" height="4">
-                    <use xlinkHref="#icon-arrow-select"></use>
-                  </svg>
-                </span>
-                <ul className="places__options places__options--custom places__options--opened">
-                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-                  <li className="places__option" tabIndex={0}>Price: low to high</li>
-                  <li className="places__option" tabIndex={0}>Price: high to low</li>
-                  <li className="places__option" tabIndex={0}>Top rated first</li>
-                </ul>
-              </form>
-              <div className="cities__places-list places__list tabs__content">
-                <article className="cities__card place-card">
+    <>
+      <article className="cities__card place-card">
                   <div className="place-card__mark">
                     <span>Premium</span>
                   </div>
@@ -98,8 +26,7 @@ export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
                     </div>
                     <div className="place-card__rating rating">
                       <div className="place-card__stars rating__stars">
-                        <span style = {{width: '80%'}}>
-                        </span>
+                        <span style={{width: '80%'}}></span>
                         <span className="visually-hidden">Rating</span>
                       </div>
                     </div>
@@ -131,8 +58,7 @@ export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
                     </div>
                     <div className="place-card__rating rating">
                       <div className="place-card__stars rating__stars">
-                        <span style= {{width: '80%'}}>
-                        </span>
+                        <span style={{width: '80%'}}></span>
                         <span className="visually-hidden">Rating</span>
                       </div>
                     </div>
@@ -164,8 +90,7 @@ export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
                     </div>
                     <div className="place-card__rating rating">
                       <div className="place-card__stars rating__stars">
-                        <span style = {{width: '80%'}}>
-                        </span>
+                        <span style={{width: '80%'}}></span>
                         <span className="visually-hidden">Rating</span>
                       </div>
                     </div>
@@ -200,8 +125,7 @@ export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
                     </div>
                     <div className="place-card__rating rating">
                       <div className="place-card__stars rating__stars">
-                        <span style={{width: '100%'}}>
-                        </span>
+                        <span style={{width: '100%'}}></span>
                         <span className="visually-hidden">Rating</span>
                       </div>
                     </div>
@@ -233,8 +157,7 @@ export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
                     </div>
                     <div className="place-card__rating rating">
                       <div className="place-card__stars rating__stars">
-                        <span style={{width: '80%'}}>
-                        </span>
+                        <span style={{width: '80%'}}></span>
                         <span className="visually-hidden">Rating</span>
                       </div>
                     </div>
@@ -244,34 +167,7 @@ export default function Main ({countArenda} : CountArendaProps) : JSX.Element {
                     <p className="place-card__type">Room</p>
                   </div>
                 </article>
-              </div>
-            </section>
-            <div className="cities__right-section">
-              <section className="cities__map map"></section>
-            </div>
-          </div>
-        </div>
-      </main>
-    </div>
+
+    </>
   );
 }
-
-/*
-function ChoiceDirection() {
-  return (
-    <>
-      {directions.map((item) => {
-        <li className="locations__item">
-          <a className="locations__item-link tabs__item" href="#">
-            <span>{item}</span>
-          </a>
-        </li>
-      }
-      )
-      };
-    </>
-  )
-};
-*/
-
-
